Allow usage of buildDocs in private repo
