+++ package.json
@@ -156,6 +156,7 @@
     "*.d.ts",
     "!scripts",
     "scripts/prReleaseNotesCommon.js",
+    "scripts/buildDocsCommon.js",
     "src",
     "testkit",
     "ReactNativeUILib.podspec",
